+++ bgstally/windows/cmdrs.py
@@ -5,6 +5,7 @@ from tkinter import ttk
 
 from bgstally.constants import DATETIME_FORMAT_JOURNAL, DiscordChannel, FONT_HEADING
 from bgstally.debug import Debug
+from bgstally.widgets import TreeviewPlus
 from ttkHyperlinkLabel import HyperlinkLabel
 
 DATETIME_FORMAT_CMDRLIST = "%Y-%m-%d %H:%M:%S"
@@ -18,7 +19,8 @@ class WindowCMDRs:
     def __init__(self, bgstally):
         self.bgstally = bgstally
 
-        self.selected_cmdr = None
+        self.selected_cmdr:dict = None
+        self.target_data:list = None
         self.toplevel:tk.Toplevel = None
 
 
@@ -52,9 +54,9 @@ class WindowCMDRs:
                         {'title': "Ship", 'type': "name", 'align': tk.W, 'stretch': tk.YES, 'width': 200},
                         {'title': "Legal", 'type': "name", 'align': tk.W, 'stretch': tk.NO, 'width': 60},
                         {'title': "Date / Time", 'type': "datetime", 'align': tk.CENTER, 'stretch': tk.NO, 'width': 150}]
-        target_data = self.bgstally.target_log.get_targetlog()
+        self.target_data = self.bgstally.target_log.get_targetlog()
 
-        treeview = TreeviewPlus(list_frame, columns=[d['title'] for d in column_info], show="headings", callback=self._cmdr_selected)
+        treeview = TreeviewPlus(list_frame, columns=[d['title'] for d in column_info], show="headings", callback=self._cmdr_selected, datetime_format=DATETIME_FORMAT_CMDRLIST)
         vsb = tk.Scrollbar(list_frame, orient=tk.VERTICAL, command=treeview.yview)
         vsb.pack(fill=tk.Y, side=tk.RIGHT)
         treeview.configure(yscrollcommand=vsb.set)
@@ -79,17 +81,22 @@ class WindowCMDRs:
             treeview.heading(column['title'], text=column['title'].title(), sort_by=column['type'])
             treeview.column(column['title'], anchor=column['align'], stretch=column['stretch'], width=column['width'])
 
-        for target in reversed(target_data):
+        for target in reversed(self.target_data):
             target_values = [target['TargetName'], target['System'], target['SquadronID'], target.get('ShipLocalised', target['Ship']), target['LegalStatus'], datetime.strptime(target['Timestamp'], DATETIME_FORMAT_JOURNAL).strftime(DATETIME_FORMAT_CMDRLIST)]
-            treeview.insert("", 'end', values=target_values)
+            iid:str = treeview.insert("", 'end', values=target_values)
+            target['iid'] = iid
 
-        if self.bgstally.discord.is_webhook_valid(DiscordChannel.BGS):
-            self.post_button =ttk.Button(buttons_frame, text="Post to Discord", command=partial(self._post_to_discord))
+        if self.bgstally.discord.is_webhook_valid(DiscordChannel.CMDR_INFORMATION) or self.bgstally.discord.is_webhook_valid(DiscordChannel.BGS):
+            self.post_button = tk.Button(buttons_frame, text="Post to Discord", command=partial(self._post_to_discord))
             self.post_button.pack(side=tk.RIGHT, padx=5, pady=5)
             self.post_button['state'] = tk.DISABLED
 
+        self.delete_button = tk.Button(buttons_frame, text="Delete", bg="red", fg="white", command=partial(self._delete_selected, treeview))
+        self.delete_button.pack(side=tk.RIGHT, padx=5, pady=5)
+        self.delete_button['state'] = tk.DISABLED
 
-    def _cmdr_selected(self, values, column):
+
+    def _cmdr_selected(self, values, column, treeview:TreeviewPlus):
         """
         A CMDR row has been clicked in the list, show details
         """
@@ -100,7 +107,13 @@ class WindowCMDRs:
 
         # Fetch the latest info for this CMDR
         self.selected_cmdr = self.bgstally.target_log.get_target_info(values[0])
-        if not self.selected_cmdr: return
+        if not self.selected_cmdr:
+            self.post_button['state'] = tk.DISABLED
+            self.delete_button['state'] = tk.DISABLED
+            return
+        else:
+            self.post_button['state'] = tk.NORMAL
+            self.delete_button['state'] = tk.NORMAL
 
         if 'TargetName' in self.selected_cmdr: self.cmdr_details_name.config(text = self.selected_cmdr['TargetName'])
         if 'inaraURL' in self.selected_cmdr: self.cmdr_details_name_inara.configure(text = "Inara Info Available", url = self.selected_cmdr['inaraURL'])
@@ -111,7 +124,18 @@ class WindowCMDRs:
         elif 'SquadronID' in self.selected_cmdr:
             self.cmdr_details_squadron.config(text = f"{self.selected_cmdr['SquadronID']}")
 
-        self.post_button['state'] = tk.NORMAL
+
+    def _delete_selected(self, treeview:TreeviewPlus):
+        """
+        Delete the currently selected CMDRs
+        """
+        selected_items:list = treeview.selection()
+        for selected_iid in selected_items:
+           for i in range(len(self.target_data)):
+               if self.target_data[i]['iid'] == selected_iid:
+                   self.target_data.pop(i) # Remove the corresponding item
+                   break
+           treeview.delete(selected_iid)
 
 
     def _post_to_discord(self):
@@ -169,51 +193,6 @@ class WindowCMDRs:
                     "inline": True
                     })
 
-        self.bgstally.discord.post_embed(f"CMDR {self.selected_cmdr['TargetName']} Spotted", None, embed_fields, None, DiscordChannel.BGS, None)
-
-
-class TreeviewPlus(ttk.Treeview):
-    def __init__(self, parent, callback, *args, **kwargs):
-        ttk.Treeview.__init__(self, parent, *args, **kwargs)
-        self.callback = callback
-        self.bind('<ButtonRelease-1>', self._select_item)
-
-
-    def heading(self, column, sort_by=None, **kwargs):
-        if sort_by and not hasattr(kwargs, 'command'):
-            func = getattr(self, f"_sort_by_{sort_by}", None)
-            if func:
-                kwargs['command'] = partial(func, column, False)
-
-        return super().heading(column, **kwargs)
-
-    def _select_item(self, event):
-        clicked_item = self.item(self.focus())
-        clicked_column_ref = self.identify_column(event.x)
-        if type(clicked_item['values']) is not list: return
-
-        clicked_column = int(clicked_column_ref[1:]) - 1
-        if clicked_column < 0: return
-
-        self.callback(clicked_item['values'], clicked_column)
-
-    def _sort(self, column, reverse, data_type, callback):
-        l = [(self.set(k, column), k) for k in self.get_children('')]
-        l.sort(key=lambda t: data_type(t[0]), reverse=reverse)
-        for index, (_, k) in enumerate(l):
-            self.move(k, '', index)
-
-        self.heading(column, command=partial(callback, column, not reverse))
-
-    def _sort_by_num(self, column, reverse):
-        self._sort(column, reverse, int, self._sort_by_num)
-
-    def _sort_by_name(self, column, reverse):
-        self._sort(column, reverse, str, self._sort_by_name)
-
-    def _sort_by_datetime(self, column, reverse):
-        def _str_to_datetime(string):
-            return datetime.strptime(string, DATETIME_FORMAT_CMDRLIST)
-
-        self._sort(column, reverse, _str_to_datetime, self._sort_by_datetime)
-
+        discord_channel:DiscordChannel = DiscordChannel.BGS
+        if self.bgstally.discord.is_webhook_valid(DiscordChannel.CMDR_INFORMATION): discord_channel = DiscordChannel.CMDR_INFORMATION
+        self.bgstally.discord.post_embed(f"CMDR {self.selected_cmdr['TargetName']} Spotted", None, embed_fields, None, discord_channel, None)